requirements.txt
Django==1.8.5
djangorestframework==3.2.4
MySQL-python==1.2.5
Pillow==3.3.1
django-webpack-loader==0.3.3
requests==2.11.1
django-cors-headers
